+++ tt/router.go
@@ -36,7 +36,7 @@ func (r *Router) Route(ctx context.Context, p mq.Packet) error {
 	return ctx.Err()
 }
 
-func (r *Router) Add(filter string, handlers ...mq.Handler) error {
+func (r *Router) Add(filter string, handlers ...mq.PubHandler) error {
 	r.routes = append(r.routes, NewRoute(filter, handlers...))
 	return nil
 }